+++ Assets/GravityDirectionPack/Scripts/GravityDirectionSystem.cs
@@ -13,16 +13,19 @@ namespace GravityDirectionPack.Scripts
         private CharacterController[] _controllers;
 
         // Start is called before the first frame update
-        void Start()
+        private void Start()
         {
-            _controllers = GameObject.FindObjectsOfType<CharacterController>();
+            ReloadControllersList();
         }
 
         // Update is called once per frame
-        void Update()
+        private void Update()
         {
             foreach (CharacterController controller in _controllers)
             {
+                if (controller == null)
+                    continue;
+
                 if (!controller.grounded)
                 {
                     controller.fallingSpeed += gravityAcceleration * Time.deltaTime;
@@ -36,8 +39,13 @@ namespace GravityDirectionPack.Scripts
                 Vector3 dir = Vector3.down;
                 Vector3 movement = dir * controller.fallingSpeed;
 
-                controller.Move(movement);
+                controller.Move(movement, Space.Self);
             }
+        }
+
+        public void ReloadControllersList()
+        {
+            _controllers = GameObject.FindObjectsOfType<CharacterController>();
         }
     }
 }